Refactor and adjust the structure of dirs.
