--- src/components/sec.vue
@@ -1,26 +1,0 @@
-<template>
-  <div id="app">
-    <router-link to="/hollow">hollow</router-link>
-  </div>
-</template>
-
-<script>
-
-export default {
-  components: {
-  },
-  mounted(){
-    console.log(this.$store.state.count)
-  }
-}
-</script>
-
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-}
-</style>
